requirements.txt
numpy
requests
python-dotenv
python-environ
redis
colored
python-dateutil
progressbar2
nltk
h5py